cp:  - changed unsigned to uint64_t for flags;  - added custom_f_*() methods

+++ core_10x/bflags.h
@@ -4,18 +4,19 @@
 #pragma once
 
 #include <sstream>
+#include <stdint.h>
 
 class BFlags {
-    unsigned        m_flags;
+    uint64_t        m_flags;
     //std::string     m_name;
 
 public:
-    static unsigned modify(unsigned state, unsigned flags_to_set, unsigned flags_to_reset)
+    static uint64_t modify(uint64_t state, uint64_t flags_to_set, uint64_t flags_to_reset)
     {
         return (state | flags_to_set) & ~flags_to_reset;
     }
 
-    static bool check(unsigned state, unsigned flags_set, unsigned flags_reset)
+    static bool check(uint64_t state, uint64_t flags_set, uint64_t flags_reset)
     {
         bool res = state & flags_set;
         if (!flags_reset)
@@ -26,9 +27,9 @@ public:
 
     BFlags() : BFlags(0x0)                                              {}
     BFlags(const BFlags& flags) = default;
-    explicit BFlags(unsigned v) : m_flags(v)                            {}
+    explicit BFlags(uint64_t v) : m_flags(v)                            {}
 
-    [[nodiscard]] unsigned value() const                                { return m_flags; }
+    [[nodiscard]] uint64_t value() const                                { return m_flags; }
     [[nodiscard]] bool on(const BFlags& flags) const                    { return m_flags & flags.m_flags; }
     [[nodiscard]] bool off(const BFlags& flags) const                   { return (m_flags & flags.m_flags) == 0x0; }
 
@@ -37,9 +38,9 @@ public:
         return check(m_flags, flags_on.m_flags, flags_off.m_flags);
     }
 
-    void set(unsigned flags)                                            { m_flags |= flags; }
-    void reset(unsigned flags)                                          { m_flags &= ~flags; }
-    void set_reset(unsigned to_set, unsigned to_reset)                  { m_flags = (m_flags | to_set) & ~to_reset; }
+    void set(uint64_t flags)                                            { m_flags |= flags; }
+    void reset(uint64_t flags)                                          { m_flags &= ~flags; }
+    void set_reset(uint64_t to_set, uint64_t to_reset)                  { m_flags = (m_flags | to_set) & ~to_reset; }
 
     void next()                                                         { m_flags <<= 1; }
